merge : feat/goodsLikeBE into back

See merge request s06-bigdata-rec-sub2/S06P22E202!217

+++ backend/backend-java/src/main/java/com/ssafy/api/service/GoodsServiceImpl.java
@@ -84,7 +84,7 @@ public class GoodsServiceImpl implements GoodsService {
     }
 
     @Override
-    public UserCommentRes goodsCommentUpdate(UserCommentReq userCommentReq, int commentSeq) {
+    public UserCommentRes goodsCommentUpdate(UserCommentReq userCommentReq, long commentSeq) {
         goodsReviewRepository.updateComment(userCommentReq.getContents(), commentSeq);
 
         UserCommentRes userCommentRes = new UserCommentRes();
